+++ tensorrt_llm/_torch/modules/fused_moe/ops/__init__.py
@@ -1,0 +1,17 @@
+# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""MoE ops module for different computation implementations."""
+
+from .moe_op import MoEOp, MoEOpSelector
+from .moe_op_cutlass import CutlassMoEOp
+from .moe_op_deepgemm import DeepGemmMoEOp
+
+__all__ = ['MoEOp', 'MoEOpSelector', 'CutlassMoEOp', 'DeepGemmMoEOp']
